refactor: move patch methods to PatchTools
